requirements.txt
opencv-python
numpy
scipy
scikit-learn
matplotlib
keras
tkinter
pillow
pandas